Update Fourier.cpp
This branch contains a version where there upper momentum boundary is determined by the number if 2pi/z wavelegths I want to go. It also includes the feature where I can force a change in approximation from the tabulated table to an approximation of vacuum being the point at which I did leave the table.