Update Fourier.cpp
The changes that I added here are to allow some exploration of how the spectral function effects the correlator without having to actually calculate the integrals for one. This should be much fast and allow some learning about the properties of the correlator and how they relate to the spectral function. The effects are going to be examined and then the branch discarded.