Update Spectral.h
Inputting a temperature dependant potential and changing the energy dependance of the spectral function to be exponential on approach to 0GeV. It may be a bit strong on that count. It may exceed the ability of the data type long double to represent just how small it is while being different from 0GeV. We're attempting to get closer to what lQCD had to say for its spatial correlation function.