Update Spectral.h
Excessive subdivision in k integral and a block that should prevent excessive subdivision in theta integral due to unknown global error. Part of the problem in 343 is that a global error estimate is not known. What I've done here is stopped integration if it becomes apparent that this part doesn't effect the result too much.