Update to Spectral.cpp and Spectral.h
This update includes a significant change to the propagators and the self-energy. The ImG_12 propagtor is now a folding integral propagtor similar to Eq.6 in NJP13. ReG_12 is a dispersion relation for BbS reduction found in Brockmann and Machleidt. The energy depenance of the self-energy has peak that depends on momentum, sqrt(M^2+P^2).