Update to Fourier.cpp
Extension of Fourier so that it can handle non-constant spectral functions when given the energy dependance of the parameterization of a+b*P^n. a(E),b(E),n(E) and their derivatives are given at points of invariant mass so that the cubic interpolation can be used inbetween those values.