Updates to Spectral.cpp and Spectral.h
This is a change from Gamma*sqrt(omega^2-k^2) to Gamma*M as the width of the quark spectrum. I don't think it completely does what Dr. Rapp wants as I'm not integrating over all energy. We think the limit is much more like the Cauchy to Dirac delta that we are looking for.