Update Spectral.cpp and Spectral.h
Passing official updates and corrections to the git. There is a change in the theta integration where I evaluate outside edges first and the Kahan summation algorithm was added to preserve significance. There is a minus sign correction in the omega width estimation for the poles in the media. There seems to be something in this copy that I don't have on the cluster as I didn't get what I was expecting at all.