Update Spectral.cpp and Spectral.h
Correct an error introduced with tracking the potential exit of the time-like region in the k-integral. Correct an error introduced with the Monte Carlo minimization where the total width of the peak integrals in k and omega didn't change with the extreme boundaries in those integrals. Currently set to calculate the table.