Update Spectral.cpp and Spectral.h
Update theta integral to have dynamic choice of intervals based on s and P. Change appears to be positive below threshold if P<20 GeV would be so kind as to agree. Above threshold is a little more questionable. Program is in Monte Carlo mode. Not a particularly good choice when the domain of boundaries are static.