Updates to Spectral.cpp Spectral.h Elements.h
Changes to these files are to correct errors that prevented the media from being calculated to the best that I can for the limits as they are provided. At this point I'm at E^1.97, better than E^4.9, but not yet at the E that Mathematica and the theory says I should be at.