Updates to Elements.h, Spectral.cpp, and Spectral.h
These changes add support for negative energy and space-like quarks in media. The peak finding algorithm has been changed out for a more robust Newton's Method. The dead reckoning method navigating peaks in the integrand has been replaced with list of limits of integration that are sorted with mergeSort. Numerically, it everything appears to be in order.