Updates to Spectral.h, Spectral.cpp, and Elements.h
These updates are mostly to Spectral.h. It changes the ReG part to be a dispersion relation deep within the integral. As this takes a forever, I've changed P=0 to be an analytic integral and so a full run on P=0 isn't awful, but the theta integral wasn't the most complicated so gains, while effective, aren't complete. Spectral.cpp adds potential parameters for Gaussian potential. Elements.h has an alteration to comparison operators to avoid examining ReG convergence at higher integral levels.