Updates to Derive.m, Derive2.m , Dispersion.cpp, Spectral.cpp, XMLScript*, and adding XMLScript2 and XMLScript5
Changes deal with covering s<0, stitching together 3 interpolation tables, and increasing coverage density for relevant s<0 while ditching coverage that is unfruitful I say that XMLScript 2 and 5 are added because they were the ones inserted in the middle of things to cover an extra table. GIT thinks XMLScript 5 and 6 are new as they didn't have those names when I started.