Update Spectral.h
Undo Shuai's suggestion. The self-energy must line up with the quark mass. The quantities tagged as quark masses should move in response to the change in quark mass. The difference is between a linear universal shift and a shift that changes to zero effect at high momentums. Effectively, by shifting the omega by the mass difference, you exacerbate the problem that you're trying to solve at high momentums rather then letting it go away as the quark energy becomes proportional to momentum and not proportional to momentum+shift.