Update Spectral.h
Updates to ImD so as make the unit make sense. This has bothered me for a long time. I really should have said something many moons back. I'm still not sure everything where it belongs or how everything comes to be this way. But this seems to be it by looking at NJP13 eq. 4