Update to Spectral.cpp and Spectral.h
This undoes a revert that was done incorrectly and reverted Spectral.cpp and not Spectral.h and then does the reversion on Spectral.h like I wanted.
The alterations to the self-energy, while possibly useful, were minor alterations that radically screwed with the Euclidean correlation function ratio. So I'm reverting to the last known good self-energy.