Update to Spectral.cpp and Spectral.h
I removed a search for the maximum of the theta integral. After some manipulation, I found an anayltic solution to its location.
There are some interesting side tangents recorded here that will be removed or igorned. The momentum dependences of the parameters is currently ignored but among the things that were changed.