Updates to Spectral.cpp and Spectral.h
Spectral.cpp has a change in destination that will need to be changed back.
Spectral.h has a number of changes intended to optimise the execution. A lot of lines were eliminated. But most of the speed increase came from the elimination of the bubble sorts yeilding ~27x improved speed.