Update to Optimiser.cpp
Locks the grid search to whatever the grid was searching rather than having it change to the best of the minimizer which can throw off the grid search. Add an API call so that it can be called by different threads and not have race conditions on the output file.