Updates to Optimiser.cpp and adding Spectral_Inter.h and Spectral_Non.h
Ditching API calls. Adding the least-squares deviation from linear of the parameters from temp to temp to the chi-squared. Looking to bind each to the next so they make sense as a series. Looking to deal with a 56-d minimization one temprature at a time rather than all at once. The minimizer is likely to get lost in that space.