Bring Optimiser.cpp from commit 993ae0dcc8 to head in new program
I brought a copy of Optimiser.cpp forward for recycled use. The current version of Optimiser.cpp will find a local minima (and preferably the global) minima for the case of linear parameter or Delta parameters. Optimiser T.cpp will then find the local minima for each T and hopefully, the parameters and delta parameters will still be reasonably linear.