Update to both Spectral.cpp and Real Contributions Spectral.h
Removed code that I'm not going to use for the enumeration of points beyond s=552.25 GeV^2.
Changed how the dispersion works so that it doesn't do zero length intervals. Zero length intervals on s'=s causes 0/0 and can be problematic as there is more than one way for s'=s to end up in the list of sub-intervals.