Update to Interpolation.h, Propagator.cpp, and Spectral.cpp, may have update for Spectral.h
Spectral.cpp now supports the interpolation of the real propagator
Output name correction in Propagtor.cpp
Adding default constructor and assignment opperator for Interpolation.h. That requires a ready flag. Wrapper made for i_k so that it can call the correct i_k depending on invariant mass square
Spectral.h is not listed in the commit manifest. If it is here: Spectral.h is in flux to meet or beat Mathematica but now supports Propagator.cpp and Spectral.cpp seperatly propducing an interpolation for Propagator and using an interpolation for the propagtor in Spectral.cpp