Update to Propagator.cpp
Terminates calculation if k > stop+50 or negative. The inverse function of i_k should not query such ranges, either because they are out of bounds or because they much larger than called for. This is to prevent exccessive execute times as k>stop+50 takes forever to get the simultanous on-shell where it matters.