Update to Propagator.cu and Spectral.cuh
The GPU code is just a bit slower than the CPU code. Breaking the GPU kernel helped make it go faster. It produces wrong results. THe program is CPU bound and the GPU doesn't even notice it in GPU utilization. The problem appears to be the overhead of launching kernels.