Update to Propagator.cpp and some changes to Propagator.cu and Spectral.cuh
The update is to convert Propagator.cpp from outputing *.m to *.csv. The *.m was such a half way thing and Mathematica wasn't really happy with it. Conversion should fix that.
The changes to Propagator.cu and Spectral.cuh was an experiment to see if the CPU could be idled. It can be, but only the there are more contexts than logical processors or if a flag is set. Otherwise, it will spinlock until the GPU finishes. I tried to set the flag, but the evidence doesn't say that it was done. I might be able to fix that with multithreading.