Updates to Elements.h, Spectral.cpp, and Spectral.h and adding Region.h to the Imaginary contributions
This commit is to save a version where I have an adaptive alogirthm for the 3d numerical integration of the imaginary contributions. I was not able to get the Lorentz variance below 40% relative change. Since that wasn't happening, I didn't do the debugging called for to get the answers near correct either. This is here in case I want to come back to it at a later time.