Update to GPU code.
The BbS reduction is corrected. I found the Folding integral and Dispersion relation to be incorrect for finite temprature. It is unknown what happened as I'm pretty sure it was working. I may just have to check out previous versions for correctness. It also seems to running than CPU code alone.